requirements.txt
langchain==0.2.1
langchain_experimental==0.0.59
python-dotenv==1.0.1
streamlit==1.35.0
mysql-connector-python
pymysql
sentence-transformers
chromadb
